Add support for registration of OpenStack domains
Domains are registered by name, and with a description. Also adds support
for the use of domains other than Default. Default is a special case, which
may be used by name. All other domains must be used by UUID, so we add
support for specifying domains by name, and perform the UUID lookup
internally.