Fix domain ID lookup for authenticating user's domain
We perform a name->ID lookup to allow domains to be specified by name. We can't
use the os_keystone_domain_facts module because ansible sanitises variables
matching anything found in the auth parameter of os_* modules. This will
include the name of the domain used to authenticate against. Use the openstack
CLI directly instead.